Removed XZ compression, refactored a bit.
